Web service application that uses three web services API
This is an integrated web service where 3 APIs have been consumed to
create an application. I have created an application which will search
for food recipe and display the list of food recipes matching the user
keywords. It also displays the nearby restaurants serving that food and
displays the location of that restaurant on google maps